add finding ahci controller

+++ arch/x86_64/inlineasm.c
@@ -14,7 +14,7 @@ void outl(uint16_t port, uint32_t val){
     __asm__ volatile ("outl %0, %1" : : "a" (val), "dN"(port));
 }
 
-uint8_t inl(uint16_t port){
+uint32_t inl(uint16_t port){
     uint32_t res;
     __asm__ volatile("inl %1, %0" : "=a" (res) : "dN" (port));
     return res;
