+++ include/arch/x86_64/inlineasm.h
@@ -4,7 +4,7 @@
 void outb(uint16_t port, uint8_t val);
 uint8_t inb(uint16_t port);
 void outl(uint16_t port, uint32_t val);
-uint8_t inl(uint16_t port);
+uint32_t inl(uint16_t port);
 void invalidate(uint64_t vaddr);
 uint64_t read_cr3(void);
 void write_cr3(uint64_t paddr);
